requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
crispy-bootstrap5==2024.10
cryptography==44.0.0
cssselect2==0.7.0
Django==5.1.3
django-allauth==65.3.1
django-crispy-forms==2.3
fonttools==4.54.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
lxml==5.3.0
oscrypto==1.3.0
packaging==24.2
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyMySQL==1.1.1
pypdf==5.1.0
pyphen==0.15.0
python-bidi==0.6.3
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.17.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tzlocal==5.2
uritools==4.0.3
urllib3==2.3.0
weasyprint==63.0
webencodings==0.5.1
wkhtmltopdf==0.2
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
